move code related to EDF+ header fields to separate module
